requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
future==0.16.0
idna==2.6
oauthlib==2.0.6
python-twitter==3.3
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
